Rename "external" to "apis"
