+++ src/containers/BitbucketKeyManager.jsx
@@ -1,6 +1,6 @@
 import {Component} from 'react';
 import {connect} from 'react-redux';
-import {setAccessToken} from '../redux/external/actions';
+import {setAccessToken} from '../redux/apis/actions';
 import {getAccessTokenFromURL} from '../utils/bitbucket';
